backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.24
langchain-community==0.4.1
langchain-google-genai==4.2.1
langgraph==1.1.6
pypdf==6.9.2
pandas==3.0.2
tabulate==0.9.0
faiss-cpu==1.13.2
sentence-transformers==5.3.0
python-dotenv==1.2.2
google-generativeai==0.8.6
streamlit==1.56.0
requests==2.33.1
